requirements.txt
evdev==1.7.1
keyboard==0.13.5
lxml==5.3.0
MouseInfo==0.1.3
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-docx==1.1.2
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
six==1.17.0
typing_extensions==4.12.2
